Feat[awt]: move AWT handler to separate package
