+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/awt/AWTInputBridge.java
@@ -1,4 +1,4 @@
-package net.kdt.pojavlaunch;
+package net.kdt.pojavlaunch.awt;
 
 public class AWTInputBridge {
     public static final int EVENT_TYPE_CHAR = 1000;
