+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/awt/AWTInputEvent.java
@@ -1,4 +1,4 @@
-package net.kdt.pojavlaunch;
+package net.kdt.pojavlaunch.awt;
 
 /*
      * Copyright (c) 1996, 2009, Oracle and/or its affiliates. All rights reserved.
