+++ app_pojavlauncher/src/main/res/layout/activity_java_gui_launcher.xml
@@ -22,7 +22,7 @@
 			tools:ignore="TouchTargetSizeCheck"
 			/>
 
-		<net.kdt.pojavlaunch.AWTCanvasView
+		<net.kdt.pojavlaunch.awt.AWTCanvasView
 			android:id="@+id/installmod_surfaceview"
 			android:layout_width="match_parent"
 			android:layout_height="match_parent"
